A1/requirements.txt
opencv==3.4.2
scikit-learn==0.24.2
numpy==1.19.2
tensorflow-gpu==1.8.0
keras==2.1.6